requirements.txt
fastapi==0.109.2
fastapi-utils==0.2.1
uvicorn[standard]==0.27.1
httpx==0.25.1
pymongo[srv]==4.6.1
python-multipart==0.0.9
passlib==1.7.4
pydantic==1.10.14
pyjwt==2.8.0

py3-validate-email==1.0.5
geopy==2.4.1

numpy==1.26.4